Pull request #9: Features dialog
Merge in INT/craiova-internship from features-dialog to firebase-test

Squashed commit of the following:

commit 737f1d26cb37efe13b7a95ecd2af2d620461ab1c
Merge: 5ab698f 15f961b
Author: Dana Bayan <dana.bayan@nagarro.com>
Date:   Fri Jul 28 10:56:31 2023 +0300

    merge features-dialog to firebase-test

commit 5ab698f3a3091893af98eec5d712d1b4f2e286c1
Author: Dana Bayan <dana.bayan@nagarro.com>
Date:   Thu Jul 27 18:43:21 2023 +0300

    added filterig on specialities and services

commit 2db8d33cffabd479db157a7408522131ce799a2b
Author: Cristian Bejan <cristian.bejan@nagarro.com>
Date:   Thu Jul 27 17:17:15 2023 +0300

    add dialog on services and specialties components

+++ src/app/core/modules/admin/components/admin-services/admin-services.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { ServicesService } from 'src/app/core/services/services.service';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { Services } from 'src/app/core/interfaces/services.interface';
+import { ConfirmationDialogService } from 'src/app/core/services/confirmation-dialog.service';
+import { MatTableDataSource } from '@angular/material/table';
 
 @Component({
   selector: 'app-admin-services',
@@ -16,22 +18,64 @@ import { Services } from 'src/app/core/interfaces/services.interface';
   ],
 })
 export class AdminServicesComponent {
+  dataSource!: MatTableDataSource<Services>;
   columnsToDisplay = ['name', 'price', 'id'];
   searchInput = '';
   columnsToDisplayWithExpand = [...this.columnsToDisplay, 'expand'];
   expandedElement!: Services;
   services: Services[] = [];
 
-  constructor(private dataBase: ServicesService) {
+  constructor(
+    private dataBase: ServicesService,
+    private dialogService: ConfirmationDialogService
+  ) {
     this.dataBase.getServices().subscribe({
       next: services => {
         this.services = services as Services[];
+        this.dataSource = new MatTableDataSource(this.services);
+        this.dataSource.filterPredicate = (data: Services, filter: string): boolean => {
+          const dataStr = Object.keys(data)
+            .reduce((currentTerm: string, key: string) => {
+              return currentTerm + (data as { [key: string]: any })[key] + '◬';
+            }, '')
+            .normalize('NFD')
+            .replace(/[\u0300-\u036f]/g, '')
+            .toLowerCase();
+
+          const transformedFilter = filter
+            .trim()
+            .normalize('NFD')
+            .replace(/[\u0300-\u036f]/g, '')
+            .toLowerCase();
+
+          return dataStr.indexOf(transformedFilter) != -1;
+        };
       },
       error: error => alert('Data could not be loaded,check your internet conection!'),
     });
   }
 
-  deleteService(id: string) {
-    this.dataBase.deleteService(id);
+  confirmDeleteDialog(id: string, name: string) {
+    const options = {
+      title: 'Stergere serviciu',
+      message: `Esti sigur ca vrei sa stergi serviciul ${name}?`,
+      cancelText: 'Nu',
+      confirmText: 'Da',
+    };
+
+    this.dialogService.open(options);
+
+    this.dialogService.confirmed().subscribe(confirmed => {
+      if (confirmed) {
+        if (id) {
+          this.dataBase.deleteService(id);
+        }
+      }
+    });
+  }
+
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value.trim().toLowerCase();
+    this.dataSource.filter = filterValue;
   }
 }
