+++ src/app/core/modules/admin/components/admin-specialties/admin-specialties.component.ts
@@ -1,7 +1,9 @@
-import { Component, Input } from '@angular/core';
+import { Component } from '@angular/core';
 import { SpecialtiesService } from 'src/app/core/services/specialties.service';
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { Specialty } from 'src/app/core/interfaces/specialty.interface';
+import { ConfirmationDialogService } from 'src/app/core/services/confirmation-dialog.service';
+import { MatTableDataSource } from '@angular/material/table';
 
 @Component({
   selector: 'app-admin-specialties',
@@ -16,22 +18,65 @@ import { Specialty } from 'src/app/core/interfaces/specialty.interface';
   ],
 })
 export class AdminSpecialtiesComponent {
+  specialties: Specialty[] = [];
+
+  dataSource!: MatTableDataSource<Specialty>;
   columnsToDisplay = ['name', 'doctorIds', 'id'];
   searchInput = '';
   columnsToDisplayWithExpand = [...this.columnsToDisplay, 'expand'];
   expandedElement!: Specialty;
-  specialties: Specialty[] = [];
 
-  constructor(private dataBase: SpecialtiesService) {
+  constructor(
+    private dataBase: SpecialtiesService,
+    private dialogService: ConfirmationDialogService
+  ) {
     this.dataBase.getSpecialties().subscribe({
       next: specialties => {
         this.specialties = specialties as Specialty[];
+        this.dataSource = new MatTableDataSource(this.specialties);
+        this.dataSource.filterPredicate = (data: Specialty, filter: string): boolean => {
+          const dataStr = Object.keys(data)
+            .reduce((currentTerm: string, key: string) => {
+              return currentTerm + (data as { [key: string]: any })[key] + '◬';
+            }, '')
+            .normalize('NFD')
+            .replace(/[\u0300-\u036f]/g, '')
+            .toLowerCase();
+
+          const transformedFilter = filter
+            .trim()
+            .normalize('NFD')
+            .replace(/[\u0300-\u036f]/g, '')
+            .toLowerCase();
+
+          return dataStr.indexOf(transformedFilter) != -1;
+        };
       },
       error: error => alert('Data could not be loaded,check your internet conection!'),
     });
   }
 
-  deleteSpecialty(id: string) {
-    this.dataBase.deleteSpecialty(id);
+  confirmDeleteDialog(id: string, name: string) {
+    const options = {
+      title: 'Stergere specializare',
+      message: `Esti sigur ca vrei sa stergi specializarea ${name}?`,
+      cancelText: 'Nu',
+      confirmText: 'Da',
+    };
+
+    this.dialogService.open(options);
+
+    this.dialogService.confirmed().subscribe(confirmed => {
+      if (confirmed) {
+        if (id) {
+          this.dataBase.deleteSpecialty(id);
+        }
+      }
+    });
+  }
+
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value.trim().toLowerCase();
+    this.dataSource.filter = filterValue;
   }
 }
